hecho, se puede mejorar

+++ JPA-SPRING/src/main/java/org/example/jpaspring/ui/RestArticle.java
@@ -36,7 +36,7 @@ public class RestArticle {
 
     @DeleteMapping("/articles/{articleId}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteArticle(@PathVariable int articleId) {
-        articleService.deleteArticle(articleId);
+    public void deleteArticle(@PathVariable int articleId, boolean confirm) {
+        articleService.deleteArticle(articleId,confirm);
     }
 }
